add better code-structure
